# crates/genomehubs-api/target/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GenomeHubs API",
    "description": "GenomeHubs Application Programming Interface.\n\nProvides a low-level API to retrieve records, look up taxa, search and generate reports across all taxon, assembly, and sample data.\n\nSee the [GenomeHubs documentation](https://genomehubs.org) for further details.",
    "contact": {
      "name": "GenomeHubs",
      "url": "https://genomehubs.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/genomehubs/genomehubs/blob/main/LICENSE"
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v3/search": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Fetch records matching a search query",
        "description": "Returns records matching a query. Supports all major query parameters including filtering, sorting, and pagination.",
        "operationId": "post_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              },
              "examples": {
                "Large mammal assemblies": {
                  "summary": "Search assemblies for mammals with genome >= 1 Gbp",
                  "value": {
                    "params_yaml": "size: 10\nfields: assembly_span,assembly_level\ntaxonomy: ncbi\n",
                    "query_yaml": "index: assembly\ntaxa:\n  - Mammalia\ntaxon_filter_type: tree\nfields:\n  - name: assembly_span\n    modifiers: [min=1000000000]\n"
                  }
                },
                "Mammalia with genome size": {
                  "summary": "Search for Mammalia taxa with a genome size estimate",
                  "value": {
                    "params_yaml": "size: 10\nfields: genome_size\ninclude_estimates: true\ntaxonomy: ncbi\n",
                    "query_yaml": "index: taxon\ntaxa:\n  - Mammalia\ntaxon_filter_type: tree\nfields:\n  - name: genome_size\n"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiStatus": {
        "type": "object",
        "description": "Uniform status block present in every v3 API response.\n\nMetadata-only endpoints (taxonomies, ranks, indices) omit `hits` and `took`.\nQuery endpoints always populate all four fields.",
        "required": [
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "hits": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "took": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "SearchRequest": {
        "type": "object",
        "required": [
          "query_yaml",
          "params_yaml"
        ],
        "properties": {
          "params_yaml": {
            "type": "string"
          },
          "query_yaml": {
            "type": "string"
          }
        }
      },
      "SearchResponse": {
        "type": "object",
        "required": [
          "status",
          "url",
          "results"
        ],
        "properties": {
          "lineage_summary": {
            "description": "Per-rank ancestor aggregation results.\n\nShape: `{rank: {ancestor_taxon_id: {field: distribution}}}`.\nOnly present when `lineage_rank_summary` was requested."
          },
          "results": {
            "type": "array",
            "items": {},
            "description": "Flat result records."
          },
          "search_after": {
            "description": "Cursor for the next page, if more results exist."
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          },
          "url": {
            "type": "string",
            "description": "URL that was built for this query (for debugging/reproduction)."
          }
        }
      },
...
    }
  },
...
}